AlgorithmWorld/requirements.txt
sqlalchemy==2.0.7
flask-sqlalchemy==2.4.1
pymysql==0.9.2
flask-jwt==0.3.2